Refactor ShardingColumnsExtractor from SPI to class
